test(gateway): extract session registry tests
